Initial Check-In
Created a new project for the SMTPSinkGo server. It's a simple server
that lets developers test outgoing email of their software without
actually sending them to people, since most people get angered by
garbage emails. It looks the email contents to the developer can confirm
the emails are correct.

This checkin includes an MIT license file, a README file with the
MarkDown format, and a starter Go file with a simple server that listens
on the SMTP port and drops connections.